--- RoundUITextField.swift
@@ -1,40 +1,0 @@
-//
-//  RoundUITextField.swift
-//  
-//
-//  Created by 임현규 on 12/08/2019.
-//
-
-import UIKit
-
-@IBDesignable
-class RoundUITextField: UITextField {
-
-  @IBInspectable var borderWidth: CGFloat {
-    get {
-      return layer.borderWidth
-    }
-    set {
-      layer.borderWidth = newValue
-    }
-  }
-  
-  @IBInspectable var cornerRadius: CGFloat {
-    get {
-      return layer.cornerRadius
-    }
-    set {
-      layer.cornerRadius = newValue
-    }
-  }
-  
-  @IBInspectable var borderColor: UIColor? {
-    get {
-      return UIColor(cgColor: layer.borderColor!)
-    }
-    set {
-      layer.borderColor = newValue?.cgColor
-    }
-  }
-  
-}
